Updating references to moved files (#62)
