+++ src/components/Map/Map.js
@@ -1,7 +1,7 @@
 import React, { Component, Props } from 'react';
 import { withScriptjs, withGoogleMap, GoogleMap, Marker, InfoWindow } from 'react-google-maps';
 import { MarkerClusterer } from 'react-google-maps/lib/components/addons/MarkerClusterer';
-import OrganizationMarker from './OrganizationMarker';
+import OrganizationMarker from './../OrganizationMarker';
 import styles from './Map.module.css';
 
 const googleMapKey = 'AIzaSyAwKdrqS2GfCt9b2K1wAopDc9Ga0N1BVUM';
